pyproject.toml
[project]
name = "aoc"
version = "0.1.0"
description = "Advent of Code utilities, solutions, clojure and python"
readme = "README.md"
authors = [
    { name = "Bartev Vartanian", email = "bartev@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "black>=24.10.0",
    "flake8>=7.1.1",
    "gnureadline>=8.2.13",
    "imageio>=2.36.1",
    "isort>=5.13.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "numpy>=2.1.3",
    "pillow>=11.0.0",
    "pip>=24.3.1",
    "prompt-toolkit>=3.0.48",
    "pydantic>=2.10.3",
    "pylint>=3.3.2",
    "pyright[all]>=1.1.389",
    "pytest>=8.3.4",
    "python-lsp-server>=1.12.0",
    "requests>=2.32.3",
    "rich[jupyter]>=13.9.4",
    "ruff>=0.8.1",
    "tqdm>=4.67.1",
    "typer>=0.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pylint]
disable = [
  "W1203", # Ignore logging-fstring-interpolation
  "C0114", # Disable docstring warnings module
  "C0115", # Disable docstring warnings class
  "C0116", # Disable docstring warnings function
  "R1732", # Disable warning for not specifying encoding
]

[tool.flake8]
ignore = [
  "E501", # Line too long
]

[dependency-groups]
dev = [
    "debugpy>=1.8.11",
    "rich>=13.9.4",
]
